Url helper and anchor tag helper tests
