+++ src/AspNeCoreSubdomain.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.16
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreSubdomain", "AspNetCoreSubdomain\AspNetCoreSubdomain.csproj", "{DB9BC0A9-320D-4B10-9D0F-5121C7C72915}"
 EndProject
@@ -16,7 +16,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WebSites", "WebSites", "{E7
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AspNetCoreSubdomain.FunctionalTests", "AspNetCoreSubdomain.FunctionalTests", "{3E2A961B-B4E8-4755-8D87-30A6F0B876EA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreSubdomain.Tests", "AspNetCoreSubdomain.Tests\AspNetCoreSubdomain.Tests.csproj", "{22C555F7-8429-48E8-82BE-30296C74B3B9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreSubdomain.Tests", "..\test\AspNetCoreSubdomain.Tests\AspNetCoreSubdomain.Tests.csproj", "{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestHelpers", "..\test\TestHelpers\TestHelpers.csproj", "{1624E56A-7F44-4CA7-8303-CB5830C06129}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -76,18 +78,30 @@ Global
 		{5796972E-7B3A-4316-B988-7734E7D9EB3F}.Release|x64.Build.0 = Release|Any CPU
 		{5796972E-7B3A-4316-B988-7734E7D9EB3F}.Release|x86.ActiveCfg = Release|Any CPU
 		{5796972E-7B3A-4316-B988-7734E7D9EB3F}.Release|x86.Build.0 = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|x64.Build.0 = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Debug|x86.Build.0 = Debug|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|x64.ActiveCfg = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|x64.Build.0 = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|x86.ActiveCfg = Release|Any CPU
-		{22C555F7-8429-48E8-82BE-30296C74B3B9}.Release|x86.Build.0 = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|x64.Build.0 = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Debug|x86.Build.0 = Debug|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|Any CPU.Build.0 = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|x64.ActiveCfg = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|x64.Build.0 = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|x86.ActiveCfg = Release|Any CPU
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462}.Release|x86.Build.0 = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|x64.Build.0 = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Debug|x86.Build.0 = Debug|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|x64.ActiveCfg = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|x64.Build.0 = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|x86.ActiveCfg = Release|Any CPU
+		{1624E56A-7F44-4CA7-8303-CB5830C06129}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -97,6 +111,10 @@ Global
 		{5796972E-7B3A-4316-B988-7734E7D9EB3F} = {E79DBCAA-D421-49DE-B290-3FDCFF2511E0}
 		{E79DBCAA-D421-49DE-B290-3FDCFF2511E0} = {FA1E1CD4-B59F-47AA-B2AE-500EC6F53C03}
 		{3E2A961B-B4E8-4755-8D87-30A6F0B876EA} = {FA1E1CD4-B59F-47AA-B2AE-500EC6F53C03}
-		{22C555F7-8429-48E8-82BE-30296C74B3B9} = {FA1E1CD4-B59F-47AA-B2AE-500EC6F53C03}
+		{53FC0B6C-4B93-47E8-A5A0-03EF9F5DF462} = {FA1E1CD4-B59F-47AA-B2AE-500EC6F53C03}
+		{1624E56A-7F44-4CA7-8303-CB5830C06129} = {FA1E1CD4-B59F-47AA-B2AE-500EC6F53C03}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6F22379D-AAD0-4C4F-9F62-E1151E33E839}
 	EndGlobalSection
 EndGlobal
