+++ test/AspNetCoreSubdomain.Tests/AspNetCoreSubdomain.Tests.csproj
@@ -15,7 +15,8 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\AspNetCoreSubdomain\AspNetCoreSubdomain.csproj" />
+    <ProjectReference Include="..\..\src\AspNetCoreSubdomain\AspNetCoreSubdomain.csproj" />
+    <ProjectReference Include="..\TestHelpers\TestHelpers.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
